The_Prisoner_Dilemma/requirements.txt
python-telegram-bot
python-statemachine
redis
redis-om
